feat(banking): first approach to banking kata
This is a first approach to this kata.

I went beyond because I didn't know what was the purpose of the Account
class. I knew I had to preserve the signature. However, this is for only
one use and didn't see how this is applicable to a real interface.
That's the reason I created the Banking class so in the bank can have
many accounts with different account ids.

This implementation supports more than one account.